+++ xds/internal/balancer/cdsbalancer/clusterHandler.go
@@ -1,145 +1,189 @@
 package cdsbalancer
+import (
+	"errors"
+	"strings"
+	"sync"
 
-import xdsclient "google.golang.org/grpc/xds/internal/client"
+	xdsclient "google.golang.org/grpc/xds/internal/client"
+)
 
-// A global to mock out in tests? To return certain functions?
-var xdsClient xdsClientInterface
 
-// Global mutex here? This should allow you to protect whole tree?
-
-// Type ClusterHandler is a wrapper around the tree of cluster(s) that represent the cluster(s)
-// being watched for in cdsbalancer. This serves as the interface for how cds balancer turns a
-// cluster name into a list of resource names to pass down to xds_cluster_resolver.
-// HAVE THE CLUSTER HANDLER SEND BACK SOMETHING, BUT HAVE THE CDS BALANCER ACTUALLY CREATE THE EDS BALANCER FROM WHAT YOU PASS IT FROM HERE
-// BLACK BOX ONTO THIS RETURNING A LIST
-// HANDLE CLUSTER HANDLER UPDATE
-/*type ClusterHandler struct {
+// ClusterHandler will be given a name representing a cluster. It will then update the CDS policy
+// constantly with a list of Clusters to pass down to XdsClusterResolverLoadBalancingPolicyConfig
+// in a stream like fashion.
+type ClusterHandler struct {
 	root *ClusterNode
-	// Or do I put the variable here?
+
+	// A way to get pinged about any updates or errors to a Node in the tree.
+	updateChannel chan string
+
+	// A mutex to protect entire tree.
+	clusterMutex sync.Mutex
+	// A channel to communicate with the cds balancer about any updates.
+	cdsUpdate chan<- []xdsclient.ClusterUpdate
+	// A channel to communicate with the cds balancer about any errors.
+	errChan chan error
+
+	xdsClient      xdsClientInterface
 }
-// I think it should be instaintiated on cds build, then expose a client conn update, which will send back a config
-// also have a channel for nodes to communicate something changed on an update, then send an updated list back by putting on a buffer
-// somewhere or something
-// handle ClientConn update
 
-// Checks for if there is a difference between the cluster name received and the cluster name of the root node (initally should be empty)
-// If there is a difference, delete node and replace it with new one?
-
-// Logically, this will take in a cluster name and return a list of child policies, every time it gets a new cluster name, or I guess every time update
-
-// Any update in the tree tell the cds balancer there is something different in regards to config
-func updateCDSBalancer
-
-func () // Handle Client Conn update from higher level, should do all logic for simply the root Node
-
-// NewClusterHandler receives a clustername from cdslbpolicy, and then constructs a cluster node related to this cluster.
-// This constructed cluster will serve as the root cluster, as in the case of Aggregate Clusters there is a potential
-// for a cluster to be a tree of clusters.
-
-// Perhaps this callback calls back into cds lb with the correct information that you trivially convert to that struct thing
-func newClusterHandler(clusterName string, callback func()) *ClusterHandler { // Is this callback used for communication with cds? Seems like this should be handleWatchUpdate.
-	return &ClusterHandler{
-		root: createClusterNode(clusterName),
+func (ch *ClusterHandler) updateRootCluster(clusterName string) /*Return value - the expectation for a stream back to the CDS Policy*/ {
+	if ch.root == nil {
+		// Construct a root node on first update.
+		ch.root = CreateClusterNode(clusterName, ch.clusterMutex, ch.updateChannel, ch.xdsClient)
+		return
 	}
-	client, err := newXDSClient()
-	if err != nil {
-		//b.logger.Errorf("failed to create xds-client: %v", err)
-		return nil // Do I return an error here?
+	// Check if root cluster was changed. If it was, delete old one and start new one, if not
+	// do nothing.
+	if clusterName != ch.root.clusterUpdate.ServiceName {
+		ch.root.delete()
+		ch.root = CreateClusterNode(clusterName, ch.clusterMutex, ch.updateChannel, ch.xdsClient)
 	}
-	xdsClient = client
-	defer func() {
-		xdsClient = nil
-	}()
-	// Defer setting to nil
-}
-// Perhaps send back a list of cluster updates, and have cds handle it
-// have a function here that starts at the root and constructs an update to send back to Cds policy
-func constructClusterUpdate() {
-	// Apparently, you use a slice here instead of a list
-	var clusterUpdateList []xdsclient.ClusterUpdate
-
-	// Base case of tree traversal: a logical dns or aggregate cluster
-	// Iterative recursion here
 }
 
-// How do we pass xdsClientInterface around, we could have it as a global?
-// I kind of like the idea of having a cluster update persisted here, and constructing the cluster update from a tree traversal from ClusterHandler
-type ClusterNode struct {
+// Watches for updates on any ClusterNode comprising the tree, and sends out a response to cds balancer.
+// Will get run from cds.
+func (ch *ClusterHandler) watcher() {
+	for {
+		select{
+		case update := <-ch.updateChannel:
+			if strings.Contains(update, "error") {
+				ch.errChan<-errors.New("There was an error in a child")
+			}
+
+			// Send CDS back a list of cluster updates which will be forwarded to the xds_cluster_resolver LB policy.
+			ch.clusterMutex.Lock()
+			// On the error case here, one of the nodes in the tree has not yet received it's first update. Thus,
+			// simply ignore this updateChannel ping.
+			if clusterUpdate, err := ch.root.constructClusterUpdate(); err != nil {
+				ch.cdsUpdate <- clusterUpdate
+			}
+			ch.clusterMutex.Unlock()
+		// TODO: a wait on a context?
+		}
+	}
+}
+
+// This logically represents a cluster. This handles all the logic for starting and stopping
+// a cluster watch, handling any updates, and constructing a list recursively for the ClusterHandler.
+type ClusterNode struct { // <- Implementation
+	// A way to cancel the watch for the cluster.
 	cancelFunc func()
+
+	// A list of children, as the Node can be an aggregate Cluster.
 	children []*ClusterNode
-
-	// Map clusterName to a ClusterNode here, for fast lookups to compute a difference
-
-	// Persist a clusterUpdate here for tree traversal to look at
+	// A ClusterUpdate in order to build a list of cluster updates for CDS to send down to child
+	// XdsClusterResolverLoadBalancingPolicy.
 	clusterUpdate xdsclient.ClusterUpdate
+	// Global mutex.
+	clusterMutex sync.Mutex
+
+	// A way to communicate with the cluster handler about a cluster update, as the cluster can be updated at any time.
+	updateChannel chan string
+
+	// This boolean determines whether this Node has received the first update or not. This isn't the best practice,
+	// but this will protect a list of Cluster Updates from being constructed if a cluster in the tree has not received
+	// an update yet.
+	receivedFirstUpdate bool
+
+	xdsClient xdsClientInterface
 }
 
-// CreateClusterNode creates a cluster, and starts a watch with the xds client.
-
-func createClusterNode(clusterName string) *ClusterNode {
-	c := &ClusterNode{}
-	c.cancelFunc = xdsClient.WatchCluster(clusterName, c.handleResp()) // All the second function does is write to the wire
+// CreateClusterNode creates a cluster node from a given clusterName. This will also start the watch for that cluster.
+func CreateClusterNode(clusterName string, clusterMutex sync.Mutex, updateChannel chan string, xdsClient xdsClientInterface) *ClusterNode {
+	c := &ClusterNode{
+		clusterMutex: clusterMutex,
+		updateChannel: updateChannel,
+		xdsClient: xdsClient,
+	}
+	// Communicate with the xds client here. TODO: Mutex handling here?
+	c.cancelFunc = xdsClient.WatchCluster(clusterName, c.handleResp)
 	return c
 }
-// MOVE ALL HANDLE LOGIC HERE
-// HandleResponse handles a xds response for a particular cluster.
-// This function also updates any children state that may have changed (i.e. if the cluster
-// was previously an aggregate cluster with children, the children could change, the
-// aggregate cluster could change to a dns cluster and lose it's children.)
-// Even if this isn't in the right place, will interface with this ClusterUpdate regardless
 
-// This function also computes the way you
-// IF INITIALLY EMPTY, WHAT'S NICE ABOUT THIS IS IT WILL STILL WORK AS DELTA OF CHILDREN WOULD BE NICE
-func (c *ClusterNode) handleResp(clusterUpdate xdsclient.ClusterUpdate) {
-	// Persist a clusterUpdate for tree traversal to look at
-	c.clusterUpdate = clusterUpdate
-	// It seems like for this, you want to compute the delta in children, then update and cancel accordingly
-
-	// First thing you do up here is get the cluster update, and construct a map
-	newChildren := make(map[string]struct{})
-	// This map will be empty if type EDS or LogicalDNS, as will have no children.
-	if clusterUpdate.ClusterType == xdsclient.ClusterTypeAggregate {
-		// Iterate through ClusterUpdate here and construct a map
-		for _, childName := range clusterUpdate.PrioritizedClusterNames {
-			newChildren[childName] = struct{}{}
-		}
-	}
-	// Iterate through the children names, which should be c.children.ClusterUpdate.
-	for _, child := range(c.children) {
-		child.clusterUpdate.ServiceName
-
-		// If the child is still in the update, then sweet
-		if _, found := newChildren[child.clusterUpdate.ServiceName]; found {
-			// Delete that child from map, as that child is already in this map
-			delete(newChildren, child.clusterUpdate.ServiceName)
-		} else { // If the child is no longer in the update, then that means the child is deleted
-			// Delete the child, this will introduce concurrency issues you then need to deal with
-			child.delete()
-		}
-	}
-
-	// whatever is left over here start watches for, which will happen implicitly through the creation of the new node
-	for child, _ := range newChildren {
-		c.children = append(c.children, createClusterNode(child))
-	}
-	// Calculate the map difference, and then afterward call into delete() or createClusterNode()
-	// on the different nodes deleted or created
-	// Construct two maps here. One map representing the response of the clusters, and one map already there in
-	// memory. Or, simply iterate through the list, where if it's not in the map
-
-	// Abstract algorithm that I developed I put here
-
+// This function serves as a wrapper around recursive logic in order to grab the lock.
+func (c *ClusterNode) delete() {
+	c.clusterMutex.Lock()
+	c.deleteRecursive()
+	c.clusterMutex.Unlock()
 }
 
-// Delete cancels the watch for the Cluster and also cancels the watch of all the Cluster's children.
-func (c *ClusterNode) delete() {
-	// TODO: any other logic...mutex locks for sure
+func (c *ClusterNode) deleteRecursive() {
 	c.cancelFunc()
 	for _, child := range c.children {
 		child.delete()
 	}
 }
-// A cluster update for a the node. The base case will be a logical dns or eds cluster, recursively call downward if aggregate
-func (c *ClusterNode) constructClusterUpdate {
 
+// Construct cluster update (potentially a list of ClusterUpdates) for a node.
+func (c *ClusterNode) constructClusterUpdate() ([]xdsclient.ClusterUpdate, error) {
+	// If the cluster has not yet received an update, the cluster update is not yet ready.
+	if !c.receivedFirstUpdate {
+		return nil, errors.New("Tried to construct a cluster update on a cluster that has not received an update.")
+	}
+
+	// Base case - LogicalDNS or EDS. Both of these cluster types will be tied to a single ClusterUpdate.
+	if c.clusterUpdate.ClusterType != xdsclient.ClusterTypeAggregate {
+		return []xdsclient.ClusterUpdate{c.clusterUpdate}, nil
+	}
+
+	// If an aggregate construct a list by recursively calling down to all of it's children.
+	var childrenUpdates []xdsclient.ClusterUpdate
+	for _, child := range c.children {
+		childUpdateList, err := child.constructClusterUpdate()
+		if err != nil {
+			return nil, err
+		}
+		for _, clusterUpdate := range childUpdateList {
+			childrenUpdates = append(childrenUpdates, clusterUpdate)
+		}
+	}
+
+	return childrenUpdates, nil
 }
+
+// handleResp handles a xds response for a particular cluster. This function also
+// handles any logic with regards to any child state that may have changed.
+func (c *ClusterNode) handleResp(clusterUpdate xdsclient.ClusterUpdate, err error) /*Changes things internally, so doesn't return anything*/ {
+	c.clusterMutex.Lock()
+	if err != nil {
+		c.updateChannel<-"There was an error."
+	}
+	c.receivedFirstUpdate = true
+	c.clusterUpdate = clusterUpdate
+
+	// This variable will determine whether there was a delta with regards to this clusterupdate. If there was, at the end
+	// of the response ping ClusterHandler to send CDS Policy a new list of ClusterUpdates.
+	var delta bool
+
+	// This map will be empty if the cluster update specifies cluster is an EDS or LogicalDNS cluster, as will have no children.
+	newChildren := make(map[string]struct{})
+	if clusterUpdate.ClusterType == xdsclient.ClusterTypeAggregate {
+		for _, childName := range clusterUpdate.PrioritizedClusterNames {
+			newChildren[childName] = struct{}{}
+		}
+	}
+
+	for _, child := range(c.children) {
+		// If the child is still present in the update, then there is nothing to do for that child name in the update.
+		if _, found := newChildren[child.clusterUpdate.ServiceName]; found {
+			delete(newChildren, child.clusterUpdate.ServiceName)
+		} else { // If the child is no longer present in the update, that cluster can be deleted.
+			delta = true
+			child.delete()
+		}
+	}
+
+	// Whatever clusters are left over here from the update are all new children, so create CDS watches for those clusters.
+	for child, _ := range newChildren {
+		delta = true
+		c.children = append(c.children, CreateClusterNode(child, c.clusterMutex, c.updateChannel, c.xdsClient))
+	}
+	// If there was a change in the state of the children, ping the ClusterHandler to construct a new update to send back
+	// to CDS.
+	if delta {
+		c.updateChannel<-"There was an update"
+	}
+	c.clusterMutex.Unlock()
+}
+
